Reverted Credentials Management to separate app
