+++ app/credentials-management/.classpath
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<classpath>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
+	<classpathentry kind="src" output="target/classes" path="src/main/java"/>
+	<classpathentry kind="src" output="target/classes" path="src/main/resources"/>
+	<classpathentry kind="src" path="/core-ui"/>
+	<classpathentry kind="src" path="/core-framework"/>
+	<classpathentry kind="output" path="target/classes"/>
+</classpath>
