+++ app/credentials-management/src/main/resources/META-INF/services/org.phoebus.framework.spi.AppDescriptor
@@ -1,0 +1,1 @@
+org.phoebus.applications.credentialsmanagement.CredentialsManagementApp
