+++ app/credentials-management/src/main/resources/css/credentials-management-style.css
@@ -1,0 +1,20 @@
+.text-field-styling{
+     -fx-padding: 5px 3px 3px 3px;
+     -fx-border-insets: 5px 3px 3px 3px;
+     -fx-background-insets: 5px 3px 3px 3px;
+     -fx-border-color: #cdcdcd;
+     -fx-border-radius: 3px;
+}
+
+.table-view .table-column{
+    -fx-alignment:center;
+}
+
+.table-view .table-cell{
+    -fx-font-weight: bold;
+    -fx-font-size: 14px;
+}
+
+.button-style{
+    -fx-pref-width: 100px;
+}
