+++ app/log-configuration/pom.xml
@@ -12,10 +12,6 @@
       <artifactId>core-ui</artifactId>
       <version>4.6.6-SNAPSHOT</version>
     </dependency>
-    <dependency>
-      <groupId>org.phoebus</groupId>
-      <artifactId>core-framework</artifactId>
-      <version>4.6.6-SNAPSHOT</version>
-    </dependency>
+
   </dependencies>
 </project>
