+++ app/pom.xml
@@ -32,5 +32,6 @@
     <module>trends</module>
     <module>utility</module>
     <module>imageviewer</module>
+    <module>credentials-management</module>
   </modules>
 </project>
