+++ core/ui/src/main/java/org/phoebus/ui/statusbar/StatusBar.java
@@ -9,9 +9,9 @@ package org.phoebus.ui.statusbar;
 
 import javafx.scene.control.Button;
 import javafx.scene.image.ImageView;
+import org.phoebus.framework.spi.AppDescriptor;
 import org.phoebus.framework.workbench.ApplicationService;
 import org.phoebus.ui.Preferences;
-import org.phoebus.ui.credentials.CredentialsManagementApp;
 import org.phoebus.ui.javafx.ImageCache;
 import org.phoebus.ui.javafx.ToolbarHelper;
 
@@ -35,12 +35,17 @@ public class StatusBar extends HBox
 
         // Show User ID?
         if (Preferences.status_show_user){
-            Button button = new Button(System.getProperty("user.name"));
-            button.setGraphic(new ImageView(ImageCache.getImage(ImageCache.class, "/icons/credentials.png")));
-            button.setOnAction(e -> {
-                ApplicationService.createInstance(CredentialsManagementApp.name);
-            });
-            getChildren().add(button);
+            AppDescriptor credentialsManagement = ApplicationService.findApplication("credentials_management");
+            if(credentialsManagement != null){
+                Button button = new Button(System.getProperty("user.name"));
+                button.setGraphic(new ImageView(ImageCache.getImage(ImageCache.class, "/icons/credentials.png")));
+                button.setOnAction(e -> credentialsManagement.create());
+                getChildren().add(button);
+            }
+            else{
+                getChildren().add(new Label(System.getProperty("user.name")));
+            }
+
         }
         // Filler between standard options and additions (update, progress, ...)
         getChildren().add(ToolbarHelper.createSpring());
