+++ phoebus-product/.classpath
@@ -52,5 +52,6 @@
     <classpathentry combineaccessrules="false" kind="src" path="/app-perfmon"/>
     <classpathentry combineaccessrules="false" kind="src" path="/app-pace"/>
     <classpathentry combineaccessrules="false" kind="src" path="/app-diag"/>
+    <classpathentry combineaccessrules="false" kind="src" path="/app-credentials-management"/>
     <classpathentry kind="output" path="target/classes"/>
 </classpath>
